source/requirements.txt
django==4.*
gunicorn
django-environ
djangorestframework
requests
markdown
django-filter
django-cors-headers
psycopg2-binary
Pillow
whitenoise
tqdm
python-magic
pandas
openpyxl
numpy
